+++ enterprise/kernel/src/main/java/org/neo4j/kernel/impl/store/format/aligned/SecondaryPageCursorReadDataAdapter.java
@@ -5,25 +5,26 @@
  * This file is part of Neo4j.
  *
  * Neo4j is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
  *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * GNU Affero General Public License for more details.
  *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.kernel.impl.store.format.highlimit;
+package org.neo4j.kernel.impl.store.format.aligned;
 
 import java.io.IOException;
 
 import org.neo4j.io.pagecache.PageCursor;
 import org.neo4j.io.pagecache.PagedFile;
-import org.neo4j.kernel.impl.store.format.highlimit.Reference.DataAdapter;
+import org.neo4j.kernel.impl.store.format.aligned.BaseAlignedRecordFormat;
+import org.neo4j.kernel.impl.store.format.aligned.Reference.DataAdapter;
 
 /**
  * {@link DataAdapter} able to acquire a secondary {@link PageCursor} on potentially a different page
@@ -61,8 +62,8 @@ class SecondaryPageCursorReadDataAdapter implements DataAdapter<PageCursor>, Sec
                 // is in fact a secondary record.
                 // TODO can we do this in BaseHighLimitRecordFormat (the place where this adapter is created) instead?
                 byte secondaryHeaderByte = secondaryCursor.getByte();
-                assert (secondaryHeaderByte & BaseHighLimitRecordFormat.HEADER_BIT_RECORD_UNIT) != 0;
-                assert (secondaryHeaderByte & BaseHighLimitRecordFormat.HEADER_BIT_FIRST_RECORD_UNIT) == 0;
+                assert (secondaryHeaderByte & BaseAlignedRecordFormat.HEADER_BIT_RECORD_UNIT) != 0;
+                assert (secondaryHeaderByte & BaseAlignedRecordFormat.HEADER_BIT_FIRST_RECORD_UNIT) == 0;
                 switched = true;
             }
             return secondaryCursor.getByte();
